requirements.txt
aiofiles==23.1.0
anyio==3.6.2
cfgv==3.3.1
click==8.1.3
distlib==0.3.6
fastapi==0.95.1
filelock==3.11.0
greenlet==2.0.2
h11==0.14.0
identify==2.5.22
idna==3.4
nodeenv==1.7.0
platformdirs==3.2.0
pre-commit==3.2.2
psycopg2-binary==2.9.6
pydantic==1.10.7
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==2.0.9
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
virtualenv==20.21.0
